# deprecate `@lookup` macro in favor of `lookup` function