--- a/docs/src/dev_reference.md
+++ b/docs/src/dev_reference.md
@@ -99,7 +99,7 @@ JuliaInterpreter.interpreted_methods
 
 ```@docs
 JuliaInterpreter.eval_code
-JuliaInterpreter.@lookup
+JuliaInterpreter.lookup
 JuliaInterpreter.is_wrapper_call
 JuliaInterpreter.is_doc_expr
 JuliaInterpreter.is_global_ref